requirements.txt
python-telegram-bot[job-queue]>=20.0
openai>=1.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
